# openapi-specs/generated_api/main.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2025-11-04T21:32:34+00:00

from __future__ import annotations

from typing import List

from fastapi import FastAPI

from .models import User, UserCreate

app = FastAPI(
    title='Test API',
    version='1.0.0',
    description='Тестовый API для демонстрации генерации FastAPI',
)


@app.get('/users', response_model=List[User])
def get_users() -> List[User]:
    """
    Получить список пользователей
    """
    pass


@app.post('/users', response_model=None)
def create_user(body: UserCreate) -> None:
    """
    Создать нового пользователя
    """
    pass


@app.get('/users/{id}', response_model=User)
def get_user_by_id(id: int) -> User:
    """
    Получить пользователя по ID
    """
    pass